【撞砖警告】【修复接力倒库的时候，取消任务直接返回原轨道】
摆渡车终止原因记录

+++ socket/Tcp/FerryTcp.cs
@@ -277,7 +277,7 @@ namespace socket.tcp
             byte[] pdata = new byte[SocketConst.FERRY_STATUS_SIZE];
             Array.Copy(data, 0, pdata, 0, SocketConst.FERRY_STATUS_SIZE);
             IDevice device = mProcess.GetStatus(pdata);
-            if (device.IsUpdate || mTimer.IsTimeOutAndReset(TimerTag.DevTcpDateRefresh, (int)DevID, 5))
+            if (device.IsUpdate || mTimer.IsTimeOutAndReset(TimerTag.DevTcpDateRefresh, DevID, 5))
             {
                 SendMsg(SocketMsgTypeE.DataReiceive, SocketConnectStatusE.通信正常, device);
                 if (device.IsUpdate) _mLog.Status(true, device.ToString());
